requirements.txt
asgiref==3.4.1
astroid==2.8.5
colorama==0.4.4
Django==3.2.9
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
pylint==2.11.1
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
wrapt==1.13.3
